Moved keys and keystates into subfolders, made pressKey and releaseKey abstract
